+++ test/descriptor/async_test.dart
@@ -62,7 +62,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });
 
@@ -101,7 +101,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });