+++ test/descriptor/directory_test.dart
@@ -102,7 +102,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });
 
@@ -140,7 +140,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });
 
@@ -182,7 +182,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });
 
@@ -228,7 +228,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });