+++ test/descriptor/pattern_test.dart
@@ -62,7 +62,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });
 
@@ -82,7 +82,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });
 
@@ -109,7 +109,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });
 
@@ -142,7 +142,7 @@ void _test(message) {
     test('test 1', () {
       scheduleSandbox();
 
-      currentSchedule.onException.schedule(() {
+      currentSchedule.onComplete.schedule(() {
         errors = currentSchedule.errors;
       });